Move the new eab model into an eab package
